Rewrite everything
We will now just generate Dockerfiles for each repository
based on the contents of the repository. We thus combine
the detector and builder in one go - making it a lot simpler
to understand and extend. It also lets us to compose them
together - making it easy to have Julia stuff in a
virtualenv based container or whatever. This was particularly
hard with s2i, since it was built on the core idea of one
container being able to run one kind of application well.