select frozen env based on Python found in environment.yml
if env specifies a version of Python other than the default,
a different base env will be selected.

This should fix some issues where switching Python versions on the base env with a bunch of packages installed can fail,
and will always produce a huge amount of churn.

Future change:

- place py2 in a separate env, like we do for requirements.txt