Use pyenv to manage base python versions
pyenv allows us to install arbitrary python versions without
having to be pinned to the distro. This separates the reproducibility
concern of packages from the concerns of python versions. As a bonus,
it also expands vastly the things that can be in runtime.txt - there
is test cases included for pypy, and many others can be installed!

This now has the clear concept of a base install, which is just
notebook setup, and then the user's install, which is whatever
version is specified in runtime.txt. `python` will point to the
version specified in runtime.txt, so executing the installed
`jupyter` will require special changes to binderhub.