Remove arguments from fetch call
This is a tiny clean-up: Since `fetch()` is only called inside the
class, and two of the three arguments in that call are attributes of
`self`, there is no need to pass them as extra arguments.

To not break the API, the parameters in the method definition are now
set to a default of `None`; if these are used with their default,
they'll be set to the respective `self` attributes. Note that this
breaks behaviour where the `self` attribute is not `None`,
but `None` is explicitly passed for the respective parameter.