Add test for host file creation in edit mode
This will test whether an externally (host) created file is seen
inside the container. This means starting the container, creating a
temporary file on the host, then running a simple verification command
(`ls`) to detect the file. Once the temporary file is removed on the
host side, it should be removed in the container as well.