Move argument parsing out of Repo2Docker class
Argument parsing should only be used when calling from
the command line, and not be deeply tied into the class
itself. This makes it easier for folks to just set
traitlets on an empty class and start it, without having to
deal with passing arguments.

This breaks how people might already be using Repo2Docker as
a library, so should not be part of 0.7