Fix cache-busting when running tests locally
When you run repo2docker locally, the uid is usually set
to the current user's uid. However, when we push the content
to Docker, we invariably set the uid of the scripts we
use as part of repo2docker at 1000. This causes a cache bust
every time.

This patch uses the correct uid, and fixes cache busting. Makes
local test runs *much* faster.

Fixes #508