tweak language to make clear `exec "$@"` works
originally it was: "The last line must be ``exec "$@"`` equivalent." Seemed open to interpretation as I needed to deduce what the equivalent of that would be here. I think adding 'or' before equivalent shows that will work, but I assume other options exist? Or why note say  "The last line must be ``exec "$@"``"?