Formatting opinion
The '//' will render to '/' in the Dockerfile that allows the RUN
command to continue on a new line. I initially added a '//' within a
python string on a blank line in order to format things within the
Python file to look potentially nicer, but this is of little importance
and I have now removed it again.