Use COPY --chown to prevent extra chown layer
COPY directive always copied files into the image owned
by root. Since we wanted this to be owned by non-root users,
we needed a RUN chown command, which added a fresh layer to
the docker image for no good reason.

This fixes that, but requires docker 19.03 to work.

Ideally, we can either:

1. Pass NB_USER to the template, so we can expand that with
   jinja2 instead of as an env. This makes it work with earlier
   versions of docker
2. Dynamically determine which version of docker we are running,
   so we can conditionally use features. Me no likey

Fixes #164