Support setting appendix on Dockerfiles too
Until now, 'appendix' was ignored for repos built with
their own Dockerfile. This fixes that, and sets the appendix
on repos built with Dockerfiles too.

I want to set a REPO_URL env var for all images built with
repo2docker-action
(https://github.com/jupyterhub/repo2docker-action/issues/86)
and this is needed for that.