Remove conda buildpacks pin of r-irkernel to 1.2
If r-base is pinned, r-irkernel will resolve to a version that is
compatible with it. But if we pin r-irkernel and not r-base, the
opposite will happen and we will end up with an older version of r-base
than is supported by r-irkernel's modern versions.

I find debugging how versions resolve with conda is really tricky, so
unless we have clear principles of what the pin should be and why, I
strongly advocate we don't have it pinned here.

In this case, having r-irkernel pinned to 1.2 caused us to get stuck at
R version 4.1 instead of going to R 4.2 that is now available.