ci: adjust to and rename extra-args.yaml to test-extra-args.yaml
I misunderstood extra-args.yaml to be a file that was recognized by
repo2docker's CLI, but in reality it was just a file recognized by the
repo2docker test suite. Following that, I renamed this testing specific
file to test-extra-args.yaml to help others avoid making that mistake
and added comments to clarify that.

This commit also gathers the conda buildpacks separate readme files into
a single readme file.