fix: create src directory in any case
The original behavior was to create an src directory with the content of the
repository. The creation would happen in any case (remote or local repository).
With the filtering in place and the default to remove the .git folder, it breaks
the build as the src folder can be missing.

This patch ensures that the directory is present in the tar so the build can
continue as it did until now.