Simulate json output from docker buildx build
We still use the docker python api to push images, and to run.
So we need to set string_output to false, and simulate the json
output on buildx. Alternatively we could completely get rid of the
python API, but that's a much bigger lift (especially around registry
credentials) that I don't want us to do right now.

Without this, push progress is reported as raw bytes in the
binderhub UI